# openapi.yaml
openapi: 3.0.3
info:
  title: Guitar Shop API
  version: 1.0.0
  description: Full API for Guitar Shop backend
servers:
  - url: https://api.guitar-shop.store
paths:
...
  /user:
    get:
      summary: Get currently logged-in user info
      tags: [Users]
      responses:
        "200":
          description: User info returned successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      email:
                        type: string
                        example: user@example.com
                      first_name:
                        type: string
                        example: John
                      last_name:
                        type: string
                        example: Doe
                  hasPwd:
                    type: boolean
                    example: true
                  error:
                    type: string
                    nullable: true
        "404":
          description: User not found
        "500":
          description: Internal server error
...
  /countries:
    get:
      summary: Get all countries
      tags: [Countries]
      responses:
        "200":
          description: List of all countries
          content:
            application/json:
              schema:
                type: object
                properties:
                  countries:
                    type: array
                    items:
                      type: object
                      properties:
                        code:
                          type: string
                          example: US
                        name:
                          type: string
                          example: United States
        "500":
          description: Internal server error
  /products:
    get:
      summary: Get all products with optional filters
      tags: [Products]
      parameters:
        - name: categoryId
          in: query
          description: Filter by category ID(s)
          required: false
          schema:
            type: array
            items:
              type: integer
        - name: brandId
          in: query
          description: Filter by brand ID(s)
          required: false
          schema:
            type: array
            items:
              type: integer
        - name: minPrice
          in: query
          description: Minimum price filter
          required: false
          schema:
            type: number
        - name: maxPrice
          in: query
          description: Maximum price filter
          required: false
          schema:
            type: number
        - name: inStock
          in: query
          description: Filter only products in stock
          required: false
          schema:
            type: boolean
        - name: sort
          in: query
          description: Sort by price, "asc" or "desc"
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: Pagination page number
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          description: Items per page
          required: false
          schema:
            type: integer
        - name: search
          in: query
          description: Search term
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of filtered products with categories, brands, and total pages
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      type: object
                  categories:
                    type: array
                    items:
                      type: object
                  brands:
                    type: array
                    items:
                      type: object
                  totalPages:
                    type: integer
        "500":
          description: Internal server error
...
  /cart:
    get:
      summary: Get current user's cart and items
      tags: [Cart]
      parameters:
        - name: userId
          in: query
          description: User ID (must match logged-in user)
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cart details with items and total price
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      type: object
                  final_price:
                    type: number
                  cart_id:
                    type: integer
        "403":
          description: Forbidden (userId mismatch)
        "500":
          description: Internal server error
...
  /checkout:
    post:
      summary: Submit an order using card details
      tags: [Checkout]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cardNumber:
                  type: string
                  example: "4111111111111111"
                expiry:
                  type: string
                  example: "12/25"
                cvv:
                  type: string
                  example: "123"
      responses:
        "201":
          description: Order processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        "400":
          description: Bad request (invalid card, empty cart, or no address)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Card details are incorrect. Please check and try again."
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error"
  /orders:
    get:
      summary: Get all orders for the currently authenticated user
      tags: [Orders]
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        user_id:
                          type: integer
                          example: 1
                        total_price:
                          type: number
                          example: 299.99
                        created_at:
                          type: string
                          format: date-time
                          example: "2025-10-01T12:34:56Z"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error"
...
  /contact:
    post:
      summary: Send a message via the contact form
      tags: [Contact]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "John Doe"
                email:
                  type: string
                  format: email
                  example: "john@example.com"
                message:
                  type: string
                  example: "I would like more information about your products."
      responses:
        "200":
          description: Email sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Email sent successfully."
        "400":
          description: Missing required fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "All fields are required."
        "500":
          description: Failed to send email
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Failed to send email."
...